[#3645] Remove o.w.m.a.w.dto.QasDto o.w.m.a.w.dto.SurveyInstanceDto
And use o.w.m.a.g.c.surveyinstance.SurveyInstanceDto instead
