--- GAE/src/org/waterforpeople/mapping/app/web/dto/QasDto.java
@@ -1,62 +1,0 @@
-/*
- *  Copyright (C) 2010-2012 Stichting Akvo (Akvo Foundation)
- *
- *  This file is part of Akvo FLOW.
- *
- *  Akvo FLOW is free software: you can redistribute it and modify it under the terms of
- *  the GNU Affero General Public License (AGPL) as published by the Free Software Foundation,
- *  either version 3 of the License or any later version.
- *
- *  Akvo FLOW is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- *  without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- *  See the GNU Affero General Public License included below for more details.
- *
- *  The full license text can also be seen at <http://www.gnu.org/licenses/agpl.html>.
- */
-
-package org.waterforpeople.mapping.app.web.dto;
-
-import java.io.Serializable;
-
-/**
- * dto that can hold questionanswerstore data
- * 
- * @author Mark Westra
- */
-public class QasDto implements Serializable {
-
-    private static final long serialVersionUID = -850583183416882347L;
-
-    // questions ids
-    private String q;
-
-    // answers
-    private String a;
-
-    // question type
-    private String t;
-
-    public String getQ() {
-        return q;
-    }
-
-    public void setQ(String q) {
-        this.q = q;
-    }
-
-    public String getA() {
-        return a;
-    }
-
-    public void setA(String a) {
-        this.a = a;
-    }
-
-    public String getT() {
-        return t;
-    }
-
-    public void setT(String t) {
-        this.t = t;
-    }
-}
